Added license commit and changed from xclip to xsel to fix bug:
The bug happened when saving to clipboard, when saved and the program
would exit, it wouldn't close xclip properly so it would stay running
and the user couldn't close the shell, it would hang until the terminal
was forcefully closed.

Changing to xsel seems to have fixed the issue as when testing, the
shell could be closed without hanging and the need to forcefully close
it.