feat(cli): check if argument of option is next
Modified mFindArgument() to check if the next entry in argv/mParameters
is the right argument instead of using just std::find() to check if it's
somewhere in the array.

Before "--format rgb" or "rgb --format" would work regardless of whether
or not you input it before or after the format option as the code only
checked if rgb exists in the array.

Now only "--format rgb" will work as the code checks if rgb is next
in the array after --format. It's a small and probably unimportant
change but it makes everything cohesive and makes the CLI look and
function like a normal CLI.