chore: remove `PreLoad.cmake` and `compile_commands.json`
PreLoad.cmake forces Ninja to be used as the build generator which can
cause some problems if the user doesn't have Ninja on the system and
doesn't want to install it just to build this project.

This change removes this file so that the user can use any build
generator they want.

compile_commands.json is also removed as it does not need to be in the
root anymore.