feat(pixels): rough linux-only pixel reading impl
At this point (on Linux) pixel reading is implemented (needs some
cleanup) and the window shows color of pixel at mouse position.

Printing (or copying to clipboard for that matter) is not yet
implemented, I have to think a bit about how I'm going to approach this.

Copying to clipboard will be implemented when reading pixels on Linux
works without issues and codebase is cleaned up, as well as when printing color values in
each format to the terminal is implemented and the codebase is cleaned up.