Abstract CopyAndConvertImage
Did some work to abstract X11-specific stuff from CopyAndConvertImage and now, it should be possible to convert an image on any platform provided information in the scp::Image::Info struct.

This is still a work-in-progress and I need touch up some stuff but right now, it displays a screenshot in the window so, nice.