Update build system
Changes to the build system. I realized I did not need SDL_image at the moment and could do everything image-related using SDL.

Removed screen_capture_lite and it's policy version minimum fully as well.

Modified the cmake/setup_sdl.cmake script a bit as well. Mainly to remove SDL_image as a dependency, but a little change is in there as well.