feat(Screenshooter): move all pixel related functions to X11
I realized after finishing the Wayland implementation, that
Info and all the other functions that handle conversion of
pixel formats are only really used in the X11 implementation,
so I took it upon myself to move everything over there.

It's really pointless to have it available for both classes,
since it's only used in one of the two classes.