feat: remove sdl_image dependency
This change completely remove sdl_image as a dependency
replacing it with stb_image.

This change is primarily aimed at build times in the case
of the user not having sdl3_image installed on their system.
In the current configuration it takes a while to configure
sdl_image as it has to download all of it's submodules and
there are quite a few. stb_image can be placed directly in
the codebase which doesn't require any downloads.